+++ src/Console/Commands/InstallSwiftAuth.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Equidna\SwifthAuth\Console\Commands;
+namespace Equidna\SwiftAuth\Console\Commands;
 
 use Illuminate\Console\Command;
 
@@ -37,7 +37,7 @@ class InstallSwiftAuth extends Command
         $this->info('Iniciando instalación de SwiftAuth...');
 
         $this->call('vendor:publish', [
-            '--provider' => 'Equidna\SwifthAuth\Providers\SwiftAuthServiceProvider',
+            '--provider' => 'Equidna\\SwiftAuth\\Providers\\SwiftAuthServiceProvider',
             '--tag' => 'swift-auth:config'
         ]);
 
@@ -62,7 +62,7 @@ class InstallSwiftAuth extends Command
         $this->info('Importando migraciones...');
 
         $this->call('vendor:publish', [
-            '--provider' => 'Equidna\SwifthAuth\Providers\SwiftAuthServiceProvider',
+            '--provider' => 'Equidna\\SwiftAuth\\Providers\\SwiftAuthServiceProvider',
             '--tag' => 'swift-auth:migrations'
         ]);
 
@@ -74,13 +74,13 @@ class InstallSwiftAuth extends Command
 
         $this->info('Importando iconos...');
         $this->call('vendor:publish', [
-            '--provider' => 'Equidna\SwifthAuth\Providers\SwiftAuthServiceProvider',
+            '--provider' => 'Equidna\\SwiftAuth\\Providers\\SwiftAuthServiceProvider',
             '--tag' => 'swift-auth:icons'
         ]);
 
         $this->info('Importando modelos...');
         $this->call('vendor:publish', [
-            '--provider' => 'Equidna\SwifthAuth\Providers\SwiftAuthServiceProvider',
+            '--provider' => 'Equidna\\SwiftAuth\\Providers\\SwiftAuthServiceProvider',
             '--tag' => 'swift-auth:models'
         ]);
 
@@ -96,7 +96,7 @@ class InstallSwiftAuth extends Command
     {
         $this->info('Instalando vistas Blade...');
         $this->call('vendor:publish', [
-            '--provider' => 'Equidna\SwifthAuth\Providers\SwiftAuthServiceProvider',
+            '--provider' => 'Equidna\\SwiftAuth\\Providers\\SwiftAuthServiceProvider',
             '--tag' => 'swift-auth:views'
         ]);
     }
@@ -111,7 +111,7 @@ class InstallSwiftAuth extends Command
     {
         $this->info('Instalando vistas en React con JavaScript...');
         $this->call('vendor:publish', [
-            '--provider' => 'Equidna\SwifthAuth\Providers\SwiftAuthServiceProvider',
+            '--provider' => 'Equidna\\SwiftAuth\\Providers\\SwiftAuthServiceProvider',
             '--tag' => 'swift-auth:js-react'
         ]);
 
@@ -128,7 +128,7 @@ class InstallSwiftAuth extends Command
     {
         $this->info('Instalando vistas en React con TypeScript...');
         $this->call('vendor:publish', [
-            '--provider' => 'Equidna\SwifthAuth\Providers\SwiftAuthServiceProvider',
+            '--provider' => 'Equidna\\SwiftAuth\\Providers\\SwiftAuthServiceProvider',
             '--tag' => 'swift-auth:ts-react'
         ]);